move copy enum to configuration instead
