+++ flytekit/tools/fast_registration.py
@@ -11,13 +11,13 @@ import tarfile
 import tempfile
 import typing
 from dataclasses import dataclass
-from enum import Enum
 from typing import Optional
 
 import click
 from rich import print as rich_print
 from rich.tree import Tree
 
+from flytekit.configuration import CopyFileDetection
 from flytekit.core.context_manager import FlyteContextManager
 from flytekit.core.utils import timeit
 from flytekit.exceptions.user import FlyteDataNotFoundException
@@ -27,17 +27,6 @@ from flytekit.tools.script_mode import _filehash_update, _pathhash_update, ls_fi
 
 FAST_PREFIX = "fast"
 FAST_FILEENDING = ".tar.gz"
-
-
-class CopyFileDetection(Enum):
-    LOADED_MODULES = 1
-    ALL = 2
-    # This option's meaning will change in the future. In the future this will mean that no files should be copied
-    # (i.e. no fast registration is used). For now, both this value and setting this Enum to Python None are both
-    # valid to distinguish between users explicitly setting --copy none and not setting the flag.
-    # Currently, this is only used for register, not for package or run because run doesn't have a no-fast-register
-    # option and package is by default non-fast.
-    NO_COPY = 3
 
 
 @dataclass(frozen=True)
@@ -106,7 +95,7 @@ def fast_package(
     # This function is temporarily split into two, to support the creation of the tar file in both the old way,
     # copying the underlying items in the source dir by doing a listdir, and the new way, relying on a list of files.
     if options and (
-        options.copy_style == CopyFileDetection.LOADED_MODULES or options.copy_style == CopyFileDetection.ALL
+            options.copy_style == CopyFileDetection.LOADED_MODULES or options.copy_style == CopyFileDetection.ALL
     ):
         if options.copy_style == CopyFileDetection.LOADED_MODULES:
             # This is the 'auto' semantic by default used for pyflyte run, it only copies loaded .py files.
