+++ tests/flytekit/unit/core/image_spec/test_image_spec.py
@@ -14,8 +14,6 @@ REGISTRY_CONFIG_FILE = os.path.join(os.path.dirname(os.path.realpath(__file__)),
 
 
 def test_uilkj():
-
-
     image_spec = ImageSpec(
         name="FLYTEKIT",
         builder="dummy",
